[Refactor] 세션 관리 클래스 이름 및 디렉토리 변경
SessionDatabase를 SessionManager로 수정하고 utils 디렉토리로 이동시켰습니다.

- 세션 ID를 관리하는 역할을 더 정확하게 반영하도록 수정
